cloud-run/requirements.txt
gunicorn==20.1.0
google-generativeai==0.3.0
python-dotenv==0.19.0
Flask==2.0.1
Werkzeug==2.0.1
Pillow==8.0.0

